+++ src/lib/authService.ts
@@ -1,10 +1,7 @@
-import {
-  signInWithEmailAndPassword,
-  signOut as firebaseSignOut,
-} from "firebase/auth";
-import { auth } from "./firebase";
-import { getUserProfile } from "./userService";
+import { collection, query, where, getDocs } from "firebase/firestore";
+import { db } from "./firebase";
 import { CachedProfile } from "./appStorage";
+import { UserProfile } from "./userService";
 
 export interface LoginResult {
   userId: string;
@@ -15,18 +12,25 @@ export async function login(
   email: string,
   password: string
 ): Promise<LoginResult> {
-  const credential = await signInWithEmailAndPassword(auth, email, password);
-  const userId = credential.user.uid;
+  const usersRef = collection(db, "users");
+  const q = query(usersRef, where("email", "==", email));
+  const snapshot = await getDocs(q);
 
-  const firestoreProfile = await getUserProfile(userId);
-  if (!firestoreProfile) {
-    throw new Error("No user profile found in Firestore.");
+  if (snapshot.empty) {
+    throw { code: "auth/user-not-found" };
   }
 
-  const { password: _omit, ...profile } = firestoreProfile;
-  return { userId, profile };
+  const userDoc = snapshot.docs[0];
+  const data = userDoc.data() as UserProfile;
+
+  if (data.password !== password) {
+    throw { code: "auth/wrong-password" };
+  }
+
+  const { password: _omit, ...profile } = data;
+  return { userId: userDoc.id, profile };
 }
 
 export async function signOut(): Promise<void> {
-  await firebaseSignOut(auth);
+  // No Firebase Auth session to clear — auth is Firestore-based.
 }
